requirements.txt
google-genai
jinja2
uvicorn
httpx
python-dotenv
